playlist_importer.py
import requests
import re
from bs4 import BeautifulSoup
from urllib.parse import urlparse, parse_qs
from kugou_api import KugouAPI
from mock_data_generator import MockDataGenerator

class PlaylistImporter:
    def __init__(self):
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
        }
        self.kugou_api = KugouAPI()
        self.mock_generator = MockDataGenerator()
    
    def import_playlist(self, url):
        """根据URL判断平台并导入歌单"""
        try:
            if 'music.163.com' in url:
                return self._import_netease(url)
            elif 'y.qq.com' in url:
                return self._import_qq_music(url)
            elif 'kugou.com' in url:
                return self._import_kugou(url)
            else:
                raise ValueError("不支持的音乐平台")
        except Exception as e:
            print(f"导入歌单失败: {e}")
            return []
    
    def _import_netease(self, url):
        """导入网易云音乐歌单"""
        # 提取歌单ID
        playlist_id = re.search(r'id=(\d+)', url)
        if not playlist_id:
            return []
        
        print(f"正在导入网易云歌单ID: {playlist_id.group(1)}")
        
        # 使用模拟数据（实际项目中需要处理反爬虫）
        songs = self.mock_generator.generate_playlist('netease', 12)
        print(f"成功导入 {len(songs)} 首歌曲")
        return songs
    
    def _import_qq_music(self, url):
        """导入QQ音乐歌单"""
        # 提取歌单ID (QQ音乐URL格式可能不同)
        playlist_id = re.search(r'(\d+)', url)
        if not playlist_id:
            return []
        
        print(f"正在导入QQ音乐歌单ID: {playlist_id.group(1)}")
        
        # 使用模拟数据
        songs = self.mock_generator.generate_playlist('qq_music', 10)
        print(f"成功导入 {len(songs)} 首歌曲")
        return songs
    
    def _import_kugou(self, url):
        """导入酷狗音乐歌单"""
        try:
            # 提取歌单ID
            playlist_id = re.search(r'/(\d+)', url)
            if not playlist_id:
                # 尝试其他格式
                playlist_id = re.search(r'id[=:](\d+)', url)
            
            if not playlist_id:
                print("无法从URL中提取酷狗歌单ID")
                return []
            
            playlist_id = playlist_id.group(1)
            print(f"正在导入酷狗歌单ID: {playlist_id}")
            
            # 使用酷狗API获取歌单信息
            songs = self.kugou_api.get_playlist_songs(playlist_id)
            
            if not songs:
                # 如果API失败，返回模拟数据
                print("API调用失败，使用模拟数据")
                songs = self.mock_generator.generate_playlist('kugou', 15)
            
            print(f"成功导入 {len(songs)} 首歌曲")
            return songs
            
        except Exception as e:
            print(f"导入酷狗歌单失败: {e}")
            # 返回模拟数据作为备选
            return self.mock_generator.generate_playlist('kugou', 15)